Refactor SequencerNode into its own module
I moved the SequencerNode struct, its associated state, constants, and trait implementations from crates/runtime/pipeline/nodes/sequencer/src/lib.rs into a new dedicated module at crates/runtime/pipeline/nodes/sequencer/src/sequencer_node.rs.

The main library file for the sequencer nodes crate (lib.rs) now declares and re-exports SequencerNode from this new module.

This change improves code organization and aligns with the structure used for other nodes like IsCueActiveNode.

Note: I skipped running tests for this change due to pre-existing MSRV (Minimum Supported Rust Version) conflicts within the project's dependencies that prevent the current version of the Rust compiler (1.75.0) from successfully building and running tests. These dependency issues need to be addressed separately.
