+++ crates/runtime/pipeline/node/src/introspection.rs
@@ -109,6 +109,7 @@ node_type_name! {
         FixtureControl,
         Programmer,
         Sequencer,
+        IsCueActive,
         Group,
         GroupControl,
         Preset,
